CategoryDraft to interface. #942
